Add more testing for properties of sensu_check
This adds a few more properties to the resource under test in the fixture
cookbook along with corresponding validation of the output using the inspec
`json` resource.

Should probably add a `check_hooks` property but there may currently be a bug
with how that's handled by `sensuctl create -f` so I chose to leave it out for
now.

Signed-off-by: Sean Escriva <sean.escriva@gmail.com>