Add sensu_user resource to provide requested feature in #33
This is an early start at a design for the sensu_user resource. There's plenty
of room for questions based on the community needs.

Much refactoring is still needed.

Note that the use of a sensitive resource property bumps the minimum required
chef version to 12.14. As this version was released more than 2 years ago in
late 2016 I believe this is a safe assumption.

* https://docs.chef.io/release_notes.html#sensitive-true
* https://github.com/chef/chef/blob/master/CHANGELOG.md#v121460-2016-09-09