accept the chef license for local dev + testing
This does not set it for use outside of test kitchen and it should be safe from my understanding as this is an OSS cookbook and it'e operators can use it with a chef distribution or a community maintained distribution. If this is incorrect we can revert this and figure something else out.

Signed-off-by: Ben Abrams <me@benabrams.it>