Use a guard and not .error? method for sensu_ctl
I don't remember the reason for trying to use .error? to check the
output of the shell exec. What matters though is that the configure is
run once when needed and not every time once it's already configured.

This change switches to using a guard and a method to compare the
configured backend url.

The shortcoming of this approach is that as resource is currently
written it supports a username and password property, however we do not
have an implementation that can safely handle reconfiguring these
values so they will only work the first time the resource is used.