Run builds in travis on xenial
Switching to xenial since it's a newer base and has support for the
addons needed to run chef integration testing.

bionic doesn't yet appear to be added to the approved list of resources
for the apt addon:

* https://github.com/travis-ci/apt-source-safelist/tree/master/keys

Additionally switched package from chefdk to chef-workstation to stay up
to date with what's maintained. Same tools, different name.