Add LDAP auth integration resource
This is essentially the same behavior as #91 that added Active Directory
integration. LDAP auth has a different object type but this attempts to
reuse the same method as much as possible

It seems logical to refactor the AD resource to follow a naming
convention such as `sensu_auth_*`, since multiple external auth provider
types are supported by sensu. However, I purposefully limited the
changes made the the `active_directory` resources in this change and did
not rename them.

Note the ldap server property callback here is the same as that used in
the `sensu_asset` resource