Add OIDC auth integration resource (#115)
This adds a resource for configuring OIDC for authentication.

The OIDC [spec
attributes](https://docs.sensu.io/sensu-go/latest/operations/control-access/auth/#oidc-spec-attributes)
are sufficiently different from Active Directory or LDAP methods that it
is a different implementation but still following similar patterns to
other resources in this cookbook.

This biggest challenge I see for users is the handling of the OIDC app
registration "client secret" value. Users should however already have
some method of handling sensitive values, such as chef-vault, Hashicorp
vault or some other method that is expected to be compatible with their
chef usage to provide a dynamic configuration value of chef-client
runtime secrets.

For this reason I have noted in the example code that it is _not_
recommended to hard code secret values in resource usage.