Add LDAP auth integration resource (#114)
* Add LDAP auth integration resource

This is essentially the same behavior as #91 that added Active Directory
integration. LDAP auth has a different object type but this attempts to
reuse the same method as much as possible

It seems logical to refactor the AD resource to follow a naming
convention such as `sensu_auth_*`, since multiple external auth provider
types are supported by sensu. However, I purposefully limited the
changes made the the `active_directory` resources in this change and did
not rename them.

Note the ldap server property callback here is the same as that used in
the `sensu_asset` resource

* Additional detail on ad_servers property deprecation

Add more info to CHANGELOG

* Use deprecate and alias for ad_servers property

Using the official approach for renaming this property instead of just
alias_method.

* Remove deprecate as errors from kitchen config

Can be re-enabled in the future but we are knowingly introducing a
deprecation warning now for a property and this would cause CI to always
fail.