Add sensu_user resource to provide requested feature in #33 (#43)
* Add sensu_user resource to provide requested feature in #33

This is an early start at a design for the sensu_user resource. There's plenty
of room for questions based on the community needs.

Much refactoring is still needed.

Note that the use of a sensitive resource property bumps the minimum required
chef version to 12.14. As this version was released more than 2 years ago in
late 2016 I believe this is a safe assumption.

* https://docs.chef.io/release_notes.html#sensitive-true
* https://github.com/chef/chef/blob/master/CHANGELOG.md#v121460-2016-09-09

* Refactor, implement disable and reinstate actions

Includes inspec tests, took some time to sort out a decent way to test for
existing users using inspec resources

* Implement modify action for user resource

There is no corresponding `sensuctl group` command, groups are created on demand
when added to users. Using the set-group subcommand is simpler for now than
trying to handle cases of add-group, remove-group or remove-groups

* Refactor sensu_user for consistency

sensu now supports using a password hash and create/modify of users from
the json spec files. This makes the sensu user resource more like the
rest of the resources provided by this cookbook while avoiding writing
user passwords to disk or stdout

* Add docs for sensu_user resource